backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.0.33
pandas
scikit-learn
joblib
